Separate decorator from operator to follow other modules in  model / project
